change rate type to float64

+++ Domain/Models/CurrencyRateModel.go
@@ -7,7 +7,7 @@ import (
 type CurrencyRateModel struct {
 	basedCurrency  Domain.Currency
 	wantedCurrency Domain.Currency
-	Rate           int
+	Rate           float64
 }
 
 func (i *CurrencyRateModel) GetBasedCurrency() Domain.Currency {
@@ -18,11 +18,11 @@ func (i *CurrencyRateModel) GetWantedCurrency() Domain.Currency {
 	return i.wantedCurrency
 }
 
-func (i *CurrencyRateModel) GetRate() int {
+func (i *CurrencyRateModel) GetRate() float64 {
 	return i.Rate
 }
 
-func GenCurrencyRateDomainModel(bCurrency Domain.Currency, wCurrency Domain.Currency, rate int) *CurrencyRateModel {
+func GenCurrencyRateDomainModel(bCurrency Domain.Currency, wCurrency Domain.Currency, rate float64) *CurrencyRateModel {
 	return &CurrencyRateModel{
 		bCurrency,
 		wCurrency,
